+++ web_ui/src/features/dataset-import/components/dataset-import-to-existing-project/dataset-import-to-existing-project-dialog.component.tsx
@@ -9,8 +9,8 @@ import { OverlayTriggerState } from '@react-stately/overlays';
 import { DATASET_IMPORT_STATUSES } from '../../../../core/datasets/dataset.enum';
 import { isAnomalyDomain } from '../../../../core/projects/domains';
 import { useProject } from '../../../../pages/project-details/providers/project-provider/project-provider.component';
-import { useDatasetImportToExistingProject } from '../../../../providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
-import { matchStatus } from '../../../../providers/dataset-import-to-existing-project-provider/utils';
+import { useDatasetImportToExistingProject } from '../../providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
+import { matchStatus } from '../../providers/dataset-import-to-existing-project-provider/utils';
 import { DatasetImportDnd } from '../dataset-import-dnd/dataset-import-dnd.component';
 import { DatasetImportProgress } from '../dataset-import-progress/dataset-import-progress.component';
 import { DatasetImportToExistingProjectDialogButtons } from './dataset-import-to-existing-project-dialog-buttons.component';
