+++ web_ui/src/features/dataset-import/components/dataset-import-to-existing-project/dataset-imports.component.tsx
@@ -15,12 +15,12 @@ import { ProjectProps } from '../../../../core/projects/project.interface';
 import { NOTIFICATION_TYPE } from '../../../../notification/notification-toast/notification-type.enum';
 import { useNotification } from '../../../../notification/notification.component';
 import { useProject } from '../../../../pages/project-details/providers/project-provider/project-provider.component';
-import { useDatasetImportToExistingProject } from '../../../../providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
-import { getLabelsMap } from '../../../../providers/dataset-import-to-existing-project-provider/utils';
 import { useWorkspaceIdentifier } from '../../../../providers/workspaces-provider/use-workspace-identifier.hook';
 import { isNonEmptyArray, isNonEmptyString } from '../../../../shared/utils';
 import { useExportImportDatasetDialogStates } from '../../../dataset-export/components/export-import-dataset-dialog-provider.component';
 import { useDatasetImportQueries } from '../../hooks/use-dataset-import-queries.hook';
+import { useDatasetImportToExistingProject } from '../../providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
+import { getLabelsMap } from '../../providers/dataset-import-to-existing-project-provider/utils';
 import { DatasetImportDeletionDialog } from '../dataset-import-deletion-dialog/dataset-import-deletion-dialog.component';
 import { DatasetImportPanel } from '../dataset-import-panel/dataset-import-panel.component';
 import { DatasetImportToExistingProjectDialog } from './dataset-import-to-existing-project-dialog.component';
