+++ web_ui/src/features/dataset-import/components/dataset-import-to-new-project/dataset-import-to-new-project-dialog.component.tsx
@@ -8,9 +8,9 @@ import { DatasetImportItem } from '../../../../core/datasets/dataset.interface';
 import { getCurrentJob } from '../../../../core/datasets/utils';
 import { useFeatureFlags } from '../../../../core/feature-flags/hooks/use-feature-flags.hook';
 import { useJobs } from '../../../../core/jobs/hooks/use-jobs.hook';
-import { useDatasetImportToNewProject } from '../../../../providers/dataset-import-to-new-project-provider/dataset-import-to-new-project-provider.component';
 import { useWorkspaceIdentifier } from '../../../../providers/workspaces-provider/use-workspace-identifier.hook';
 import { isNonEmptyString, runWhenTruthy } from '../../../../shared/utils';
+import { useDatasetImportToNewProject } from '../../providers/dataset-import-to-new-project-provider/dataset-import-to-new-project-provider.component';
 import { DatasetImportDeletionDialog } from '../dataset-import-deletion-dialog/dataset-import-deletion-dialog.component';
 import { DatasetImportToNewProjectDialogButtons } from './dataset-import-to-new-project-dialog-buttons.component';
 import { DatasetImportToNewProjectDialogContent } from './dataset-import-to-new-project-dialog-content.component';
