+++ web_ui/src/features/dataset-import/providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component.tsx
@@ -9,27 +9,27 @@ import { useQueryClient } from '@tanstack/react-query';
 import { HttpStatusCode } from 'axios';
 import { isEmpty } from 'lodash-es';
 
-import { DATASET_IMPORT_STATUSES } from '../../core/datasets/dataset.enum';
-import { DatasetImportToExistingProjectItem } from '../../core/datasets/dataset.interface';
-import { ProjectIdentifier } from '../../core/projects/core.interface';
-import { CreateDatasetResponse } from '../../core/projects/dataset.interface';
-import { useDatasetImportQueries } from '../../features/dataset-import/hooks/use-dataset-import-queries.hook';
-import { useLocalStorageDatasetImport } from '../../features/dataset-import/hooks/use-local-storage-dataset-import.hook';
-import { PinnedCollapsedItemsAction } from '../../hooks/use-pinned-collapsed-items/use-pinned-collapsed-items.interface';
-import { NOTIFICATION_TYPE } from '../../notification/notification-toast/notification-type.enum';
-import { useNotification } from '../../notification/notification.component';
-import { useDatasetIdentifier } from '../../pages/annotator/hooks/use-dataset-identifier.hook';
-import { useMedia } from '../../pages/media/providers/media-provider.component';
-import { useProject } from '../../pages/project-details/providers/project-provider/project-provider.component';
-import { LOCAL_STORAGE_KEYS } from '../../shared/local-storage-keys';
-import { MissingProviderError } from '../../shared/missing-provider-error';
-import { getFileSize, runWhenTruthy } from '../../shared/utils';
-import { idMatchingFormat } from '../../test-utils/id-utils';
+import { DATASET_IMPORT_STATUSES } from '../../../../core/datasets/dataset.enum';
+import { DatasetImportToExistingProjectItem } from '../../../../core/datasets/dataset.interface';
+import { ProjectIdentifier } from '../../../../core/projects/core.interface';
+import { CreateDatasetResponse } from '../../../../core/projects/dataset.interface';
+import { PinnedCollapsedItemsAction } from '../../../../hooks/use-pinned-collapsed-items/use-pinned-collapsed-items.interface';
+import { NOTIFICATION_TYPE } from '../../../../notification/notification-toast/notification-type.enum';
+import { useNotification } from '../../../../notification/notification.component';
+import { useDatasetIdentifier } from '../../../../pages/annotator/hooks/use-dataset-identifier.hook';
+import { useMedia } from '../../../../pages/media/providers/media-provider.component';
+import { useProject } from '../../../../pages/project-details/providers/project-provider/project-provider.component';
+import { useDataset } from '../../../../providers/dataset-provider/dataset-provider.component';
+import { useTusUpload } from '../../../../providers/tus-upload-provider/tus-upload-provider.component';
+import { getUploadId, onErrorMessage, throttleProgress } from '../../../../providers/tus-upload-provider/util';
+import { useWorkspaceIdentifier } from '../../../../providers/workspaces-provider/use-workspace-identifier.hook';
+import { LOCAL_STORAGE_KEYS } from '../../../../shared/local-storage-keys';
+import { MissingProviderError } from '../../../../shared/missing-provider-error';
+import { getFileSize, runWhenTruthy } from '../../../../shared/utils';
+import { idMatchingFormat } from '../../../../test-utils/id-utils';
+import { useDatasetImportQueries } from '../../hooks/use-dataset-import-queries.hook';
+import { useLocalStorageDatasetImport } from '../../hooks/use-local-storage-dataset-import.hook';
 import { getBytesRemaining, getTimeRemaining } from '../dataset-import-to-new-project-provider/utils';
-import { useDataset } from '../dataset-provider/dataset-provider.component';
-import { useTusUpload } from '../tus-upload-provider/tus-upload-provider.component';
-import { getUploadId, onErrorMessage, throttleProgress } from '../tus-upload-provider/util';
-import { useWorkspaceIdentifier } from '../workspaces-provider/use-workspace-identifier.hook';
 import { getDatasetImportInitialState, getLabelsMap, matchStatus } from './utils';
 
 interface DatasetImportToExistingProjectContextProps {
