+++ web_ui/src/pages/create-project/create-project-button/create-project-button.test.tsx
@@ -10,7 +10,7 @@ import { AxiosError } from 'axios';
 import { useStatus } from '../../../core/status/hooks/use-status.hook';
 import { TOO_LOW_FREE_DISK_SPACE_IN_BYTES } from '../../../core/status/hooks/utils';
 import { StatusProps } from '../../../core/status/status.interface';
-import { DatasetImportToNewProjectProvider } from '../../../providers/dataset-import-to-new-project-provider/dataset-import-to-new-project-provider.component';
+import { DatasetImportToNewProjectProvider } from '../../../features/dataset-import/providers/dataset-import-to-new-project-provider/dataset-import-to-new-project-provider.component';
 import { ProjectsImportProvider } from '../../../providers/projects-import-provider/projects-import-provider.component';
 import { applicationRender } from '../../../test-utils/application-provider-render';
 import {
