+++ web_ui/src/pages/project-details/components/project-dataset/project-dataset-tab-actions.test.tsx
@@ -14,7 +14,7 @@ import {
     ExportImportDatasetDialogProvider,
     useExportImportDatasetDialogStates,
 } from '../../../../features/dataset-export/components/export-import-dataset-dialog-provider.component';
-import { DatasetImportToExistingProjectProvider } from '../../../../providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
+import { DatasetImportToExistingProjectProvider } from '../../../../features/dataset-import/providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
 import { DatasetProvider } from '../../../../providers/dataset-provider/dataset-provider.component';
 import { MediaUploadProvider } from '../../../../providers/media-upload-provider/media-upload-provider.component';
 import { applicationRender as render } from '../../../../test-utils/application-provider-render';
