+++ web_ui/src/routes/projects/dataset.route.tsx
@@ -4,8 +4,8 @@
 import { Outlet } from 'react-router-dom';
 
 import { ExportImportDatasetDialogProvider } from '../../features/dataset-export/components/export-import-dataset-dialog-provider.component';
+import { DatasetImportToExistingProjectProvider } from '../../features/dataset-import/providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
 import { MediaProvider } from '../../pages/media/providers/media-provider.component';
-import { DatasetImportToExistingProjectProvider } from '../../providers/dataset-import-to-existing-project-provider/dataset-import-to-existing-project-provider.component';
 import { DatasetProvider } from '../../providers/dataset-provider/dataset-provider.component';
 
 export const DatasetRoute = () => {
